- 서비스에 있던 방대한 로직을 도메인 및 일급 컬렉션으로 분리
- 리스트로 관리하던 엔티티를 일급 컬렉션으로 관리
- 기존에 확인하지 못한 예외 처리

+++ src/main/java/admin/adminsiteserver/common/aws/infrastructure/S3Uploader.java
@@ -49,14 +49,16 @@ public class S3Uploader {
     }
 
     public void delete(List<FilePathDto> filePathDtos) {
+        if (filePathDtos == null) return;
         for (FilePathDto filePathDto : filePathDtos) {
             delete(filePathDto);
         }
     }
 
     public void delete(FilePathDto filePathDto) {
+        if (filePathDto == null) return;
         String fileName = filePathDto.getFileUrl().replace(cloudfront, "");
-        if (amazonS3.doesObjectExist(bucketName, fileName)) {
+        if (!fileName.isBlank() && amazonS3.doesObjectExist(bucketName, fileName)) {
             amazonS3.deleteObject(new DeleteObjectRequest(bucketName, fileName));
         }
     }
